@@ -1,33 +1,41 @@
 name: nct
 description: "Novus configuration tool"
-publish_to: 'none'
+publish_to: "none"
 version: 0.1.0
 
 environment:
-  sdk: ^3.7.2
+  sdk: ">=3.8.0 <4.0.0"
 
 dependencies:
   dart_jsonwebtoken: ^3.2.0
   flutter:
     sdk: flutter
   flutter_svg: ^2.2.1
-  go_router: ^16.0.0
+  go_router: ^17.0.0
   google_fonts: ^6.2.1
   http: ^1.3.0
+  json_annotation: ^4.9.0
   network_info_plus: ^7.0.0
   provider: ^6.1.5
   reorderables: ^0.6.0
   url_launcher: ^6.3.1
   web: ^1.1.1
+  freezed_annotation: ^3.1.0
+  flutter_form_builder: ^10.2.0
+  form_builder_validators: ^11.2.0  # Add this line
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_lints: ^6.0.0
+  build_runner: ^2.5.4
+  json_serializable: ^6.7.1
+  freezed: ^3.2.3
+
 
 flutter:
   uses-material-design: true
-  
+
   assets:
     - assets/images/logo.png
     - assets/images/NovusLogo.svg
