@@ -1,232 +1,189 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:nct/pages/device/deviceOverviewPage.dart';
-import 'package:nct/pages/network/networkOverviewPage.dart';
-import 'package:nct/api/UserApi.dart';
-import 'package:nct/pages/errorPage.dart';
-import 'package:nct/explorerLayout.dart';
-import 'package:nct/pages/loginPage.dart';
 import 'package:nct/pages/logoutPage.dart';
-import 'package:nct/models.dart';
+import 'package:nct/pages/network/networkPage.dart';
+import 'package:nct/api/PublicApi.dart';
+import 'package:nct/pages/errorPage.dart';
+import 'package:nct/layout.dart';
+import 'package:nct/pages/loginPage.dart';
 import 'package:nct/pages/ntp/ntpOverviewPage.dart';
 import 'package:nct/pages/snmp/snmpOverviewPage.dart';
 import 'package:nct/pages/supportPage.dart';
-
 import 'package:nct/pages/users/usersOverviewPage.dart';
+import 'package:nct/api/TimeApi.dart';
 
-import 'package:provider/provider.dart';
-
-// Consolidated Route Configuration
-
-final loginRoute = RouteConfig(
-  path: "/login",
-  name: "Login",
-  icon: Icons.login,
-  pageBuilder: () => LoginPage(),
-  isAllowed: true,
-);
-
-final logoutRoute = RouteConfig(
-  path: "/logout",
-  name: "Logout",
-  icon: Icons.logout,
-  pageBuilder: () => LogoutPage(),
-);
-
-final usersRoute = RouteConfig(
-  path: '/users',
-  name: 'System Users',
-  icon: Icons.people,
-  pageBuilder: () => UsersOverviewPage(),
-  isAllowed: true,
-);
-
-//final deviceRoute = RouteConfig(
-//  path: '/device',
-//  name: 'Device ',
-//  icon: Icons.dns,
-//  pageBuilder: () => DeviceOverviewPage(),
-//  isAllowed: true,
-//);
-
-final snmpRoute = RouteConfig(
-  path: '/snmp',
-  name: 'Snmp',
-  icon: Icons.alarm,
-  pageBuilder: () => SnmpOverviewPage(),
-);
-
-final networkRoute = RouteConfig(
-  path: '/network',
-  name: 'Network',
-  icon: Icons.assessment,
-  pageBuilder: () => NetworkOverviewPage(),
-);
-
-final ntpRoute = RouteConfig(
-  path: '/ntp',
-  name: 'Ntp',
-  icon: Icons.assessment,
-  pageBuilder: () => NtpOverviewPage(),
-);
-
-final supportRoute = RouteConfig(
-  path: '/support',
-  name: 'Support',
-  icon: Icons.help,
-  pageBuilder: () => SupportPage(),
-  isAllowed: true,
-);
-
-// ================================================================
-// Router Configuration
 class AppRouter {
-  static final GoRouter router = GoRouter(
-    initialLocation: '/login',
-    redirect: (context, state) {
-      final authApi = Provider.of<UserApi>(context, listen: false);
-      final loggedIn = authApi.isLoggedIn;
+  static final _publicRoutes = ['/', '/login', '/support'];
 
-      final isLoginRoute = state.uri.path == '/login';
+  static GoRouter createRouter(PublicApi publicApi) {
+    return GoRouter(
+      initialLocation: '/login',
+      refreshListenable: publicApi,
+      redirect: (context, state) {
+        final path = state.uri.path;
+        final isPublic = _publicRoutes.contains(path);
 
-      // If not authenticated and trying to access protected route, redirect to login
-      if (!loggedIn && AppRoutes.requiresAuthentication(state.uri.path)) {
-        return '/login';
-      }
+        if (!publicApi.isLoggedIn && !isPublic) return '/login';
+        if (publicApi.isLoggedIn && path == '/login') return '/network';
+        return null;
+      },
+      routes: [
+        GoRoute(
+          path: '/',
+          redirect: (context, state) =>
+              publicApi.isLoggedIn ? '/network' : '/login',
+        ),
 
-      // If authenticated and on login page, redirect to dashboard
-      if (loggedIn && isLoginRoute) {
-        return '/network';
-      }
-
-      // Check if route is allowed by server
-      if (!AppRoutes.isRouteAllowed(state.uri.path)) {
-        return loggedIn ? '/network' : '/login';
-      }
-
-      return null; // Allow navigation
-    },
-    routes: [
-      GoRoute(
-        path: '/',
-        redirect: (BuildContext context, GoRouterState state) {
-          final authApi = Provider.of<UserApi>(context, listen: false);
-          return authApi.isLoggedIn ? '/network' : '/login';
-        },
-      ),
-
-      ShellRoute(
-        builder: (context, state, child) {
-          return ExplorerLayout(child: child);
-        },
-        routes: AppRoutes.generateGoRoutes(),
-      ),
-    ],
-    errorPageBuilder:
-        (context, state) =>
-            NoTransitionPage(child: ErrorPage(error: state.error.toString())),
-  );
-}
-
-// ================================================================
-// Single source of truth for all routes
-class AppRoutes {
-  static final List<RouteConfig> routes = [
-    //deviceRoute,
-    networkRoute,
-    ntpRoute,
-    loginRoute,
-    snmpRoute,
-    usersRoute,
-    supportRoute,
-    logoutRoute,
-  ];
-
-  // Utility methods to work with route configuration
-  static List<String> getAllowedPaths() {
-    List<String> paths = ['/', '/login', '/support'];
-
-    void addPaths(List<RouteConfig> configs, String parentPath) {
-      for (var config in configs) {
-        String fullPath = parentPath + config.path;
-        if (config.isAllowed) {
-          paths.add(fullPath);
-        }
-        if (config.children.isNotEmpty) {
-          addPaths(config.children, fullPath);
-        }
-      }
-    }
-
-    addPaths(routes, '');
-    return paths;
-  }
-
-  static List<GoRoute> generateGoRoutes() {
-    List<GoRoute> goRoutes = [];
-
-    GoRoute buildRoute(RouteConfig config, String parentPath) {
-      String fullPath = parentPath + config.path;
-
-      return GoRoute(
-        path: config.path,
-        pageBuilder:
-            (context, state) => NoTransitionPage(child: config.pageBuilder()),
-        routes:
-            config.children
-                .map((child) => buildRoute(child, fullPath))
-                .toList(),
-      );
-    }
-
-    for (var route in routes) {
-      goRoutes.add(buildRoute(route, ''));
-    }
-
-    return goRoutes;
+        ShellRoute(
+          builder: (context, state, child) => ExplorerLayout(child: child),
+          routes: [
+            GoRoute(
+              path: '/network',
+              pageBuilder: (context, state) =>
+                  NoTransitionPage(child: NetworkPage()),
+            ),
+            GoRoute(
+              path: '/ntp',
+              pageBuilder: (context, state) =>
+                  NoTransitionPage(child: NtpOverviewPage()),
+            ),
+            GoRoute(
+              path: '/login',
+              pageBuilder: (context, state) =>
+                  NoTransitionPage(child: LoginPage()),
+            ),
+            GoRoute(
+              path: '/snmp',
+              pageBuilder: (context, state) =>
+                  NoTransitionPage(child: SnmpOverviewPage()),
+            ),
+            GoRoute(
+              path: '/users',
+              pageBuilder: (context, state) =>
+                  NoTransitionPage(child: UsersOverviewPage()),
+            ),
+            GoRoute(
+              path: '/support',
+              pageBuilder: (context, state) =>
+                  NoTransitionPage(child: SupportPage()),
+            ),
+            GoRoute(
+              path: '/logout',
+              pageBuilder: (context, state) =>
+                  NoTransitionPage(child: LogoutPage()),
+            ),
+          ],
+        ),
+      ],
+      errorPageBuilder: (context, state) =>
+          NoTransitionPage(child: ErrorPage(error: state.error.toString())),
+    );
   }
 
   static List<ExplorerItem> generateExplorerItems(bool isLoggedIn) {
-    //inal loginApi = Provider.of<LoginApi>(context, listen: false);
-    // Only return items if user is authenticated
-    List<ExplorerItem> items = [];
-
-    ExplorerItem buildItem(RouteConfig config, String parentPath) {
-      String fullPath = parentPath + config.path;
-
-      return ExplorerItem(
-        id: fullPath.replaceAll('/', '_'),
-        name: config.name,
-        route: fullPath,
-        icon: config.icon,
-        children:
-            config.children.map((child) => buildItem(child, fullPath)).toList(),
-      );
-    }
-
     if (!isLoggedIn) {
-      items.add(buildItem(loginRoute, ''));
-      items.add(buildItem(supportRoute, ''));
-      return items;
+      return [
+        ExplorerItem(
+          id: '_login',
+          name: 'Login',
+          route: '/login',
+          icon: Icons.login,
+          children: [],
+        ),
+        ExplorerItem(
+          id: '_support',
+          name: 'Support',
+          route: '/support',
+          icon: Icons.help,
+          children: [],
+        ),
+      ];
     }
 
-    for (var route in routes) {
-      // Exclude /login for logged-in users
-      if (route.isAllowed && route.path != '/login') {
-        items.add(buildItem(route, ''));
-      }
-    }
-
-    return items;
+    return [
+      ExplorerItem(
+        id: '_network',
+        name: 'Network',
+        route: '/network',
+        icon: Icons.assessment,
+        children: [],
+      ),
+      ExplorerItem(
+        id: '_ntp',
+        name: 'Ntp',
+        route: '/ntp',
+        icon: Icons.assessment,
+        children: [],
+      ),
+      ExplorerItem(
+        id: '_snmp',
+        name: 'Snmp',
+        route: '/snmp',
+        icon: Icons.alarm,
+        children: [],
+      ),
+      ExplorerItem(
+        id: '_users',
+        name: 'Users',
+        route: '/users',
+        icon: Icons.people,
+        children: [],
+      ),
+      ExplorerItem(
+        id: '_support',
+        name: 'Support',
+        route: '/support',
+        icon: Icons.help,
+        children: [],
+      ),
+      ExplorerItem(
+        id: '_logout',
+        name: 'Logout',
+        route: '/logout',
+        icon: Icons.assessment,
+        children: [],
+      ),
+    ];
   }
+}
 
-  static bool isRouteAllowed(String route) {
-    return getAllowedPaths().contains(route);
-  }
+class RouteConfig {
+  final String path;
+  final String name;
+  final IconData icon;
+  final Widget Function() pageBuilder;
 
-  static bool requiresAuthentication(String route) {
-    // Public routes that don't require authentication
-    List<String> publicRoutes = ['/', '/login', '/support'];
-    return !publicRoutes.contains(route);
-  }
+  final List<RouteConfig> children;
+  final bool isAllowed;
+  final bool requiresAuth;
+
+  const RouteConfig({
+    required this.path,
+    required this.name,
+    required this.icon,
+    required this.pageBuilder,
+    this.children = const [],
+    this.isAllowed = true,
+    this.requiresAuth = true,
+  });
+}
+
+// Model classes
+class ExplorerItem {
+  final String id;
+  final String name;
+  final String route;
+  final IconData icon;
+  final List<ExplorerItem> children;
+  bool isExpanded;
+
+  ExplorerItem({
+    required this.id,
+    required this.name,
+    required this.route,
+    required this.icon,
+    this.children = const [],
+    this.isExpanded = false,
+  });
 }
